Move kubernetes_scripts to kubernetes
Maybe this will conflict with the kubernetes module? Anyway it seems
okay in my virtualenv...
